Add support for the Smol async runtime
Since async-std has been discontinued in favor of smol, we introduce support
for smol.

Unlike tokio and async-std, smol does not provide a mechanism to detect whether
the current task is running within its runtime. As a result, runtime selection
is determined at compile time, requiring exactly one of the "tokio", "smol", or
"async_std" features to be enabled.

Additionally, unlike tokio and async-std, smol cancels a task on drop instead of
detaching it. This behavior required adjustments in the Client code to properly
handle detached tasks.