Refactor client nodes
Client nodes are simplified and redesigned.

ConstantNode variant is removed, `.app` and `.app/device` nodes
can be implemented as static nodes.

Static nodes are implemented using async trait.

The client no longer handles application state. The generic type
denoting an app state is thus removed from the client API. It is
up to the application to capture its state and whatever it needs
in request handlers.